Added out-of-band binding for lambdas.
I want a way to uniquely refer to any function object.  That means
binding them to identifiers as they are encountered.  ctx.bindFn
allows for this.

Still needs a better way to generate the identifier names since,
with the current code, every instance of a macro-generated function
will have the same name.