Auto. /ouro/lang import now injected in correct order.
Previously, the implicit import to /ouro/lang was injected separately
to the module's statements.  However, this actually causes the second
set to be injected *after* the implicit if done whilst compiling a
statement.

Prepending the implicit statement to the start of the module's
statements ensures the correct relative order.

This didn't actually break anything; it just meant that any statements
using builtin functions or macros would have to be compiled at least
twice.